requirements.txt
streamlit
ultralytics
torch
torchvision
pillow